add conditional if before unlink for delete

+++ application/controllers/Gigs.php
@@ -140,7 +140,9 @@ class Gigs extends CI_Controller
             $this->activity_model->cancel_gig($notif);
         }
 
-        unlink('uploads/' . $gig[0]['file']);
+        if (!empty($gig[0]['file']))
+            unlink('uploads/' . $gig[0]['file']);
+
         $this->gig_model->delete_gig($id);
         $this->session->set_flashdata('flash', "The gig was successfully deleted!");
         redirect("gigs");
